+++ lib/src/db/isoexpress/rt_lmp_hourly.dart
@@ -4,29 +4,33 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:async';
 import 'package:collection/collection.dart';
+import 'package:csv/csv.dart';
 import 'package:date/date.dart';
+import 'package:duckdb_dart/duckdb_dart.dart';
+import 'package:elec/elec.dart';
 import 'package:elec_server/src/db/config.dart';
+import 'package:elec_server/src/db/lib_prod_dbs.dart';
+import 'package:logging/logging.dart';
+import 'package:path/path.dart';
 import 'package:timezone/timezone.dart';
 import '../lib_mis_reports.dart' as mis;
 import '../lib_iso_express.dart';
 import '../converters.dart';
 import 'package:dotenv/dotenv.dart' as dotenv;
 
-
 class RtLmpHourlyArchive extends DailyIsoExpressReport {
-
   RtLmpHourlyArchive({ComponentConfig? dbConfig, String? dir}) {
     dbConfig ??= ComponentConfig(
-          host: '127.0.0.1', dbName: 'isoexpress', collectionName: 'rt_lmp_hourly');
+        host: '127.0.0.1',
+        dbName: 'isoexpress',
+        collectionName: 'rt_lmp_hourly');
     this.dbConfig = dbConfig;
     dir ??= '${baseDir}PricingReports/RtLmpHourly/Raw/';
     this.dir = dir;
     reportName = 'Real-Time Energy Market Hourly LMP Report';
   }
 
-  // @override
-  // String getUrl(Date? asOfDate) =>
-  //     'https://www.iso-ne.com/static-transform/csv/histRpts/rt-lmp/lmp_rt_final_${yyyymmdd(asOfDate)}.csv';
+  static final log = Logger('ISONE RT LMP');
 
   @override
   String getUrl(Date? asOfDate) =>
@@ -64,11 +68,10 @@ class RtLmpHourlyArchive extends DailyIsoExpressReport {
     }
   }
 
-
   @override
   Map<String, dynamic> converter(List<Map<String, dynamic>> rows) {
     var row = <String, dynamic>{};
-    row['date'] = (rows.first['BeginDate'] as String).substring(0,10);
+    row['date'] = (rows.first['BeginDate'] as String).substring(0, 10);
     row['ptid'] = int.parse(rows.first['Location']['@LocId']);
     row['congestion'] = <num>[];
     row['lmp'] = <num>[];
@@ -104,10 +107,11 @@ class RtLmpHourlyArchive extends DailyIsoExpressReport {
 
   List<Map<String, dynamic>> _processFileJson(File file) {
     var aux = json.decode(file.readAsStringSync()) as Map;
-    late List<Map<String,dynamic>> xs;
+    late List<Map<String, dynamic>> xs;
     if (aux.containsKey('HourlyLmps')) {
-      if (aux['HourlyLmps'] == '') return <Map<String,dynamic>>[];
-      xs = (aux['HourlyLmps']['HourlyLmp'] as List).cast<Map<String,dynamic>>();
+      if (aux['HourlyLmps'] == '') return <Map<String, dynamic>>[];
+      xs =
+          (aux['HourlyLmps']['HourlyLmp'] as List).cast<Map<String, dynamic>>();
     } else {
       throw ArgumentError('Can\'t find key HourlyLmps in file $file');
     }
@@ -148,8 +152,8 @@ class RtLmpHourlyArchive extends DailyIsoExpressReport {
     var pwd = dotenv.env['ISONE_WS_PASSWORD']!;
 
     var client = HttpClient()
-      ..addCredentials(Uri.parse(getUrl(day)), '',
-          HttpClientBasicCredentials(user, pwd))
+      ..addCredentials(
+          Uri.parse(getUrl(day)), '', HttpClientBasicCredentials(user, pwd))
       ..userAgent = 'Mozilla/4.0'
       ..badCertificateCallback = (cert, host, port) {
         print('Bad certificate connecting to $host:$port:');
@@ -161,6 +165,127 @@ class RtLmpHourlyArchive extends DailyIsoExpressReport {
     await response.pipe(getFilename(day).openWrite());
   }
 
+  /// Aggregate a list of days into a csv.gz file.  Each row of the file
+  /// looks like:
+  ///
+  /// ```dart
+  /// {
+  ///   'ptid': 321,
+  ///   'date': '2022-12-22',
+  ///   'hour': 0,
+  ///   'extraHourDst': false,
+  ///   'lmp': <num>,
+  ///   'mcc': <num>,
+  ///   'mlc: <num>,
+  /// }
+  /// ```
+  List<Map<String, dynamic>> aggregateDays(List<Date> days) {
+    assert(days.first.location == IsoNewEngland.location);
+    var out = <Map<String, dynamic>>[];
+    for (var date in days) {
+      log.info('...  Working on $date');
+      final hours = date.hours();
+      final file = getFilename(date, extension: 'json');
+      if (file.existsSync()) {
+        var rows = processFile(file);
+        for (var row in rows) {
+          for (var i = 0; i < hours.length; i++) {
+            var extraHourDst = false;
+            var hour = hours[i];
+            if (hour.start.hour == hour.previous.start.hour) {
+              extraHourDst = true;
+            }
+            var one = {
+              'ptid': row['ptid'],
+              'date': row['date'],
+              'hour': hour.start.hour,
+              'extraHourDst': extraHourDst,
+              'lmp': row['lmp'][i],
+              'mcc': row['congestion'][i],
+              'mlc': row['marginal_loss'][i],
+            };
+            out.add(one);
+          }
+        }
+      } else {
+        throw StateError('Missing file for $date');
+      }
+    }
+    log.info('aggregated data has ${out.length} rows!');
+    return out;
+  }
+
+  int makeGzFileForMonth(Month month) {
+    var days = month.days();
+    final xs = aggregateDays(days);
+    final file = File('$dir../month/rt_lmp_${month.toIso8601String()}.csv');
+    var converter = const ListToCsvConverter();
+    var sb = StringBuffer();
+    sb.writeln(converter.convert([xs.first.keys.toList()]));
+    for (var offer in xs) {
+      sb.writeln(converter.convert([offer.values.toList()]));
+    }
+    file.writeAsStringSync(sb.toString());
+
+    // gzip it!
+    var res = Process.runSync('gzip', ['-f', file.path], workingDirectory: dir);
+    if (res.exitCode != 0) {
+      throw StateError('Gzipping ${basename(file.path)} has failed');
+    }
+    log.info('Gzipped file ${basename(file.path)}');
+
+    return 0;
+  }
+
+  /// Create the Db from scratch
+  int rebuildDuckDb() {
+    final dbPath = '$baseDir/rt_lmp.duckdb';
+    if (File(dbPath).existsSync()) File(dbPath).deleteSync();
+    final con = Connection(dbPath);
+    con.execute('''
+CREATE TABLE IF NOT EXISTS rt_lmp (
+    ptid UINTEGER NOT NULL,
+    date DATE NOT NULL,
+    hour UTINYINT NOT NULL,
+    extraDstHour BOOL NOT NULL,
+    lmp DECIMAL(9,4) NOT NULL,
+    mcc DECIMAL(9,4) NOT NULL,
+    mcl DECIMAL(9,4) NOT NULL,
+);
+''');
+    con.execute('''
+INSERT INTO rt_lmp
+FROM read_csv(
+    '${dirname(dir)}/month/rt_lmp_*.csv.gz', 
+    header = true, 
+    columns = {
+      'ptid': 'UINTEGER',
+      'date': 'DATE',
+      'hour': 'UTINYINT',
+      'extraDstHour': 'BOOL',
+      'lmp': 'DECIMAL(9,4)', 
+      'mcc': 'DECIMAL(9,4)', 
+      'mcl': 'DECIMAL(9,4)', 
+    },
+    dateformat = '%Y-%m-%d');
+''');
+
+    // add the NERC holidays for convenience
+    con.execute('''
+CREATE TABLE nerc_holidays (
+    date DATE NOT NULL,
+);
+INSERT INTO nerc_holidays
+FROM read_csv(
+    '${DuckDbProd.base}/Calendars/nerc_holidays.csv', 
+    header = true, 
+    dateformat = '%Y-%m-%d');
+''');
+
+    con.close();
+
+    return 0;
+  }
 
   /// Check if this date is in the db already
   Future<bool> hasDay(Date date) async {
@@ -183,8 +308,7 @@ class RtLmpHourlyArchive extends DailyIsoExpressReport {
           'date': 1,
         },
         unique: true);
-    await dbConfig.db
-        .createIndex(dbConfig.collectionName, keys: {'date': 1});
+    await dbConfig.db.createIndex(dbConfig.collectionName, keys: {'date': 1});
     await dbConfig.db.close();
   }